+++ src/index.ts
@@ -2,6 +2,6 @@ export type {UploadConfig, UploadedFileMetadata, UploadFileConfig} from './uploa
 
 export type {TransferredFileInfo} from './transfer';
 
-export type {ObjectiveRef} from './findOrCreate';
+export type {ObjectiveRef} from './find-or-create';
 
 export {getDefaultUploadConfig} from './upload-config';
